+++ superwall/src/main/java/com/superwall/sdk/paywall/vc/SuperwallPaywallActivity.kt
@@ -2,6 +2,8 @@ package com.superwall.sdk.paywall.vc
 
 import android.Manifest
 import android.R
+import android.animation.ArgbEvaluator
+import android.animation.ValueAnimator
 import android.app.Activity
 import android.app.NotificationChannel
 import android.app.NotificationManager
@@ -9,21 +11,26 @@ import android.content.Context
 import android.content.Intent
 import android.content.pm.PackageManager
 import android.graphics.Color
+import android.graphics.drawable.ColorDrawable
 import android.os.Build
 import android.os.Bundle
+import android.util.Log
 import android.view.View
 import android.view.ViewGroup
 import android.view.Window
 import android.view.WindowInsetsController
 import android.view.WindowManager
+import android.widget.FrameLayout
 import androidx.activity.OnBackPressedCallback
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.app.AppCompatDelegate
+import androidx.coordinatorlayout.widget.CoordinatorLayout
 import androidx.core.app.ActivityCompat
 import androidx.core.app.NotificationManagerCompat
 import androidx.core.content.ContextCompat
 import androidx.core.view.children
+import com.google.android.material.bottomsheet.BottomSheetBehavior
 import com.superwall.sdk.Superwall
 import com.superwall.sdk.dependencies.DeviceHelperFactory
 import com.superwall.sdk.misc.isLightColor
@@ -90,16 +97,72 @@ class SuperwallPaywallActivity : AppCompatActivity() {
 
     private var contentView: View? = null
     private var notificationPermissionCallback: NotificationPermissionCallback? = null
+    private val isBottomSheetView
+        get() = contentView is CoordinatorLayout && Build.VERSION.SDK_INT >= Build.VERSION_CODES.R
 
     override fun setContentView(view: View) {
         super.setContentView(view)
         contentView = view
     }
 
+    private val mainScope = CoroutineScope(Dispatchers.Main)
+
+    private fun paywallView(): PaywallView? {
+        val content = contentView ?: error("Content view is null")
+        if (content is PaywallView) {
+            return content
+        } else if (content is CoordinatorLayout) {
+            return (content.getChildAt(0) as FrameLayout).getChildAt(0) as PaywallView
+        } else {
+            return null
+        }
+    }
+
+    private fun setupBottomSheetLayout(paywallView: PaywallView) {
+        val activityView =
+            layoutInflater.inflate(com.superwall.sdk.R.layout.activity_bottom_sheet, null)
+        setContentView(activityView)
+        initBottomSheetBehavior()
+        val container =
+            activityView.findViewById<FrameLayout>(com.superwall.sdk.R.id.container)
+        container.setOnClickListener { finish() }
+        container.addView(paywallView)
+        container.requestLayout()
+    }
+
+    private fun initBottomSheetBehavior() {
+        var bottomSheetBehavior = BottomSheetBehavior.from((contentView as ViewGroup).getChildAt(0))
+        bottomSheetBehavior.halfExpandedRatio = 0.62f
+        // Expanded by default
+        bottomSheetBehavior.state = BottomSheetBehavior.STATE_HALF_EXPANDED
+        bottomSheetBehavior.skipCollapsed = true
+        bottomSheetBehavior.addBottomSheetCallback(
+            object :
+                BottomSheetBehavior.BottomSheetCallback() {
+                override fun onStateChanged(
+                    bottomSheet: View,
+                    newState: Int,
+                ) {
+                    if (newState == BottomSheetBehavior.STATE_HIDDEN) {
+                        finish()
+                    }
+                }
+
+                override fun onSlide(
+                    bottomSheet: View,
+                    slideOffset: Float,
+                ) {
+                }
+            },
+        )
+    }
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
         AppCompatDelegate.setCompatVectorFromResourcesEnabled(true)
+        val presentationStyle =
+            intent.getSerializableExtra(PRESENTATION_STYLE_KEY) as? PaywallPresentationStyle
 
         // Show content behind the status bar
         window.decorView.systemUiVisibility = View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN or
@@ -136,10 +199,16 @@ class SuperwallPaywallActivity : AppCompatActivity() {
                 return
             }
 
+        val isBottomSheetStyle = presentationStyle == PaywallPresentationStyle.DRAWER
+
         (view.parent as? ViewGroup)?.removeView(view)
         view.encapsulatingActivity = WeakReference(this)
-
-        setContentView(view)
+        // If it's a bottom sheet, we set activity as transparent and show the UI in a bottom sheet container
+        if (isBottomSheetStyle && Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
+            setupBottomSheetLayout(view)
+        } else {
+            setContentView(view)
+        }
 
         onBackPressedDispatcher.addCallback(
             this,
@@ -162,14 +231,14 @@ class SuperwallPaywallActivity : AppCompatActivity() {
         when (intent.getSerializableExtra(PRESENTATION_STYLE_KEY) as? PaywallPresentationStyle) {
             PaywallPresentationStyle.PUSH -> {
                 if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.UPSIDE_DOWN_CAKE) {
-                    overrideActivityTransition(OVERRIDE_TRANSITION_OPEN, R.anim.slide_in_left, R.anim.slide_in_left)
+                    overrideActivityTransition(
+                        OVERRIDE_TRANSITION_OPEN,
+                        R.anim.slide_in_left,
+                        R.anim.slide_in_left,
+                    )
                 } else {
                     overridePendingTransition(R.anim.slide_in_left, R.anim.slide_in_left)
                 }
-            }
-
-            PaywallPresentationStyle.DRAWER -> {
-                // TODO: Not yet supported in Android
             }
 
             PaywallPresentationStyle.FULLSCREEN -> {
@@ -189,12 +258,10 @@ class SuperwallPaywallActivity : AppCompatActivity() {
             PaywallPresentationStyle.MODAL -> {
                 // TODO: Not yet supported in Android
             }
-
-            PaywallPresentationStyle.NONE -> {
-                // Do nothing
-            }
-
-            null -> {
+            PaywallPresentationStyle.NONE,
+            PaywallPresentationStyle.DRAWER,
+            null,
+            -> {
                 // Do nothing
             }
         }
@@ -202,7 +269,7 @@ class SuperwallPaywallActivity : AppCompatActivity() {
 
     override fun onStart() {
         super.onStart()
-        val paywallVc = contentView as? PaywallView ?: return
+        val paywallVc = paywallView() ?: return
 
         if (paywallVc.isBrowserViewPresented) {
             paywallVc.isBrowserViewPresented = false
@@ -211,19 +278,55 @@ class SuperwallPaywallActivity : AppCompatActivity() {
         paywallVc.beforeViewCreated()
     }
 
+    private fun setBottomSheetTransparency() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
+            setTranslucent(true)
+            val colorFrom = Color.argb(0, 0, 0, 0)
+            val colorTo = Color.argb(200, 0, 0, 0)
+            with(ValueAnimator.ofObject(ArgbEvaluator(), colorFrom, colorTo)) {
+                setDuration(600) // milliseconds
+                addUpdateListener { animator -> window.setBackgroundDrawable(ColorDrawable(animator.animatedValue as Int)) }
+                start()
+            }
+        } else {
+            window.setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT))
+            setTheme(android.R.style.Theme_Translucent_NoTitleBar)
+        }
+    }
+
+    private fun hideBottomSheetAndFinish() {
+        val colorFrom = Color.argb(200, 0, 0, 0)
+        val colorTo = Color.argb(0, 0, 0, 0)
+
+        // First animate the background dim, then call finish on the view
+        with(ValueAnimator.ofObject(ArgbEvaluator(), colorFrom, colorTo)) {
+            setDuration(300) // milliseconds
+            addUpdateListener { animator ->
+                val e = ((animator.animatedValue as Int) / colorFrom)
+                Log.e("Percentage", " -------------------------- $e -------------")
+                if (e < 0.1) {
+                    super.finish()
+                }
+                window.setBackgroundDrawable(ColorDrawable(animator.animatedValue as Int))
+            }
+            start()
+        }
+    }
+
     override fun onResume() {
         super.onResume()
-        val paywallVc = contentView as? PaywallView ?: return
-
+        val paywallVc = paywallView() ?: return
+        if (isBottomSheetView) {
+            setBottomSheetTransparency()
+        }
         paywallVc.onViewCreated()
     }
 
     override fun onPause() {
         super.onPause()
 
-        val paywallVc = contentView as? PaywallView ?: return
-
-        CoroutineScope(Dispatchers.Main).launch {
+        val paywallVc = paywallView() ?: return
+        mainScope.launch {
             paywallVc.beforeOnDestroy()
         }
     }
@@ -231,9 +334,9 @@ class SuperwallPaywallActivity : AppCompatActivity() {
     override fun onStop() {
         super.onStop()
 
-        val paywallVc = contentView as? PaywallView ?: return
+        val paywallVc = paywallView() ?: return
 
-        CoroutineScope(Dispatchers.Main).launch {
+        mainScope.launch {
             paywallVc.destroyed()
         }
     }
@@ -242,7 +345,7 @@ class SuperwallPaywallActivity : AppCompatActivity() {
         super.onDestroy()
         (contentView?.parent as? ViewGroup)?.removeView(contentView)
         // Clear reference to activity in the view
-        (contentView as? ActivityEncapsulatable)?.encapsulatingActivity = null
+        (paywallView() as? ActivityEncapsulatable)?.encapsulatingActivity = null
 
         // Clear the reference to the contentView
         contentView = null
@@ -251,6 +354,16 @@ class SuperwallPaywallActivity : AppCompatActivity() {
     //region Notifications
     interface NotificationPermissionCallback {
         fun onPermissionResult(granted: Boolean)
+    }
+
+    override fun finish() {
+        if (isBottomSheetView) {
+            mainScope.launch {
+                hideBottomSheetAndFinish()
+            }
+        } else {
+            super.finish()
+        }
     }
 
     suspend fun attemptToScheduleNotifications(